requirements.txt
appdirs==1.4.3
asgiref==3.3.1
beautifulsoup4==4.9.3
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
coloredlogs==14.0
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
Cython==0.29.21
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-extra-fields==3.0.2
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
html5lib==1.0.1
humanfriendly==9.0
idna==2.8
image==1.5.33
img2pdf==0.4.0
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
lockfile==0.12.2
lxml==4.6.2
MarkupSafe==1.1.1
msgpack==0.6.2
ocrmypdf==11.3.4
openapi-codec==1.3.2
packaging==20.3
pdfminer.six==20201018
pep517==0.8.2
pikepdf==2.2.0
Pillow==8.0.1
pluggy==0.13.1
progress==1.5
psycopg2==2.8.6
pycparser==2.20
pyparsing==2.4.6
PyPDF2==1.26.0
pytesseract==0.3.6
pytoml==0.1.21
pytz==2020.4
reportlab==3.5.56
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scapy==2.4.4
simplejson==3.17.2
six==1.14.0
sortedcontainers==2.3.0
soupsieve==2.0.1
sqlparse==0.4.1
tinyocr==0.2.3
tqdm==4.54.1
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
